Rework and rename the ABI update script
Switch to using the existing Autonity build scripts to generate the ABI
files, as there is no need to roll our own.

The script is renamed from `update_abi` to `update-abi.sh` in line with
current project conventions.

The script now handles paths with spaces.

The use of `solc-version.txt` was dropped because there is no need for
it right now, and also it would be better for this file to be generated
by the Autonity Makefile itself.