Allow customizing RPC Endpoint URL for tests
The current tests were hard-coded to use a Piccadilly testnet RPC
endpoint. This change makes it possible to define `TEST_RPC_URL` either
as an environment variable or in a `.env` [1] file.

A `.env` pattern at the repo root is added to the `.gitignore` as this
file should be treated as a user-private configuration file and not part
of the repo itself.

[1] https://github.com/theskumar/python-dotenv