Try to make project URLs start with upper case
In the left hand side of PyPI project page, most headings are
capitalized. Whereas in the case of "Project links", they are in lower
case. Try to change the case of the `pyproject.toml` entries to see if
it makes a difference, as I could not find any guidance about this.