Cast contract function return values to tuple for parsing
Fixes `AssertionError` from the

    assert isinstance(return_value, tuple)

assertion when calling the `cdps` getter of the Stabilization contract.

The reason for failing the assertion is that for this call the contract
function wrapper returns a list instead of the expected tuple.
To work around it, cast the return value to a tuple.

Note that according to web3.py docs[1], the function wrapper's call()
method "returns the return value of the executed function" i.e. there is
no guarantee that the return value is either a list or a tuple, however
I don't think we should worry about it until we run into similar errors.

[1]: https://web3py.readthedocs.io/en/stable/web3.contract.html#web3.contract.ContractFunction.call

Fixes #37.