Add Makefile for generating contract bindings
Use the `pyabigen` tool (https://github.com/clearmatics/pyabigen)
to generate Autonity contract bindings.

Use a Makefile instead of the old shell script because it can track
which targets are up-to-date.

All contract functions that are marked with `onlyProtocol` and
`onlyAutonity` modifiers in the contract code are excluded from
the generated bindings with the `--exclude` command line argument.