Replace the ERC20 contract binding
The contract binding that has been generated for the IERC20 contract
interface from the Autonity repository[1] doesn't include the accessor
functions for `symbol`, `name` and `decimals` that are part of the ERC20
token standard and are frequently used.

Replace the binding with one generated for OpenZeppelin's ERC20
implementation[2] instead, which includes all expected functions.

Remove the rule from Makefile because it would add much complexity to
clone and build the openzeppelin-contracts repository and it is very
unlikely that the ERC20 binding will ever need to be re-generated.

[1]: https://github.com/autonity/autonity/blob/release/v0.14.1/autonity/solidity/contracts/interfaces/IERC20.sol
[2]: https://github.com/OpenZeppelin/openzeppelin-contracts/blob/v5.1.0/contracts/token/ERC20/ERC20.sol