Ignore patch releases when checking protocol version
When a contract binding is created, autonity.py checks whether the
binding matches the Autonity protocol version. This is done by comparing
the client version that the bindings were generated for with with the
client version of the RPC node, and showing a warning if the two
versions don't match.

This method of comparison may cause problems.

Firstly, it requires a new autonity.py release for every Autonity client
release even if there is no change in the contract interface.

Secondly, it is easily possible that different nodes of the network run
different patch versions of the Autonity client.

To resolve this issue, ignore the patch version when comparing version
numbers as patch releases only contain bugfixes and don't change the
contract interface.